+++ examples/v3.0/components/newtype-numeric.rs
@@ -1,0 +1,13 @@
+pub mod schemas {
+    pub struct IntegerDefaultValue(i64);
+
+    pub struct Int64Value(i64);
+
+    pub struct Int32Value(i32);
+
+    pub struct NumberDefaultValue(f64);
+
+    pub struct FloatValue(f32);
+
+    pub struct DoubleValue(f64);
+}
